move layout to layout dir update imports
