+++ pages/_app.js
@@ -1,6 +1,6 @@
 import { globalStyles } from '../shared/styles'
 import 'bootstrap/dist/css/bootstrap.min.css';
-import Layout from '../components/layout';
+import Layout from '../components/layout/layout';
 
 
 export default function MyApp({ Component, pageProps }) {
